Added an example for passing a double vector or array.
This behaves pretty much like the cantera interface for 
things like arrays of mole fractions, where you pass
a vector of doubles, instead of a character array.
You again pass a pointer to an array of zeros,
the C++ code copies into it, and you retrieve the result.
But this time, the pointer value is modified.